refactor(web): search-params に関する処理を states.ts に移動
