+++ packages/web/src/states.ts
@@ -1,6 +1,7 @@
-import { createSignal } from "solid-js";
+import { createEffect, createSignal } from "solid-js";
 import { sourceNames } from "spigot-event-list-common";
 import { createStore } from "solid-js/store";
+import { isServer } from "solid-js/web";
 
 export const [searchText, setSearchText] = createSignal("");
 
@@ -14,4 +15,42 @@ export function resetFilterSources() {
 
 function defaultFilterSources() {
   return Object.fromEntries(sourceNames.map((source) => [source, true]));
+}
+
+function initFromSearchParams() {
+  const params = new URLSearchParams(window.location.search);
+  if (params.has("search")) {
+    setSearchText(params.get("search"));
+  }
+  if (params.has("tags")) {
+    const sources = params.get("tags").split("-");
+    setFilterSources(
+      Object.fromEntries(
+        sourceNames.map((source) => [source, sources.includes(source)])
+      )
+    );
+  }
+}
+
+export function createStatesEffect() {
+  if (!isServer) {
+    initFromSearchParams();
+    createEffect(() => {
+      const paramsArray: string[][] = [];
+      const search = searchText();
+      const sources = sourceNames.filter((source) => filterSources[source]);
+      if (search) {
+        paramsArray.push(["search", search]);
+      }
+      if (sourceNames.length !== Object.keys(sources).length) {
+        paramsArray.push(["tags", sources.join("-")]);
+      }
+      const path = window.location.pathname;
+      const params = paramsArray.length
+        ? "?" + new URLSearchParams(paramsArray).toString()
+        : "";
+      const hash = window.location.hash;
+      window.history.replaceState(null, "", path + params + hash);
+    });
+  }
 }
